chore(team): split mailbox error helpers
